+++ opm/simulators/flow/countGlobalCells.hpp
@@ -24,7 +24,7 @@
 #ifndef OPM_COUNTGLOBALCELLS_HEADER_INCLUDED
 #define OPM_COUNTGLOBALCELLS_HEADER_INCLUDED
 
-#include <opm/core/props/BlackoilPhases.hpp>
+#include <opm/simulators/utils/BlackoilPhases.hpp>
 
 #include <dune/grid/common/gridview.hh>
