+++ opm/simulators/wells/GasLiftStage2.hpp
@@ -20,8 +20,9 @@
 #ifndef OPM_GASLIFT_STAGE2_HEADER_INCLUDED
 #define OPM_GASLIFT_STAGE2_HEADER_INCLUDED
 
-#include <opm/core/props/BlackoilPhases.hpp>
 #include <opm/simulators/wells/GasLiftSingleWellGeneric.hpp>
+
+#include <opm/simulators/utils/BlackoilPhases.hpp>
 
 #include <map>
 #include <memory>
