+++ opm/simulators/wells/MultisegmentWellSegments.cpp
@@ -41,7 +41,7 @@
 #include <opm/simulators/wells/SegmentState.hpp>
 #include <opm/simulators/wells/WellInterfaceGeneric.hpp>
 
-#include <opm/core/props/BlackoilPhases.hpp>
+#include <opm/simulators/utils/BlackoilPhases.hpp>
 
 #include <fmt/format.h>
