+++ opm/simulators/wells/RegionAverageCalculator.hpp
@@ -20,8 +20,9 @@
 #ifndef OPM_REGIONAVERAGECALCULATOR_HPP_HEADER_INCLUDED
 #define OPM_REGIONAVERAGECALCULATOR_HPP_HEADER_INCLUDED
 
-#include <opm/core/props/BlackoilPhases.hpp>
 #include <opm/simulators/wells/RegionAttributeHelpers.hpp>
+
+#include <opm/simulators/utils/BlackoilPhases.hpp>
 #include <opm/simulators/utils/DeferredLoggingErrorHelpers.hpp>
 
 #include <dune/grid/common/gridenums.hh>
