+++ opm/simulators/wells/StandardWellConnections.cpp
@@ -20,9 +20,8 @@
 */
 
 #include <config.h>
-#include <opm/simulators/wells/StandardWellConnections.hpp>
 
-#include <opm/core/props/BlackoilPhases.hpp>
+#include <opm/simulators/wells/StandardWellConnections.hpp>
 
 #include <opm/material/fluidsystems/BlackOilFluidSystem.hpp>
 
@@ -30,7 +29,9 @@
 #include <opm/models/blackoil/blackoilonephaseindices.hh>
 #include <opm/models/blackoil/blackoiltwophaseindices.hh>
 
+#include <opm/simulators/utils/BlackoilPhases.hpp>
 #include <opm/simulators/utils/DeferredLoggingErrorHelpers.hpp>
+
 #include <opm/simulators/wells/ParallelWellInfo.hpp>
 #include <opm/simulators/wells/WellInterfaceIndices.hpp>
 #include <opm/simulators/wells/WellState.hpp>
