+++ tests/test_relpermdiagnostics.cpp
@@ -23,8 +23,10 @@
 
 #define BOOST_TEST_MODULE RelpermDiagnostics
 
+#include <opm/simulators/utils/satfunc/RelpermDiagnostics.hpp>
 
 #include <opm/common/utility/platform_dependent/disable_warnings.h>
+
 #include <boost/test/unit_test.hpp>
 #include <boost/version.hpp>
 #if BOOST_VERSION / 100000 == 1 && BOOST_VERSION / 100 % 1000 < 71
@@ -32,18 +34,21 @@
 #else
 #include <boost/test/tools/floating_point_comparison.hpp>
 #endif
+
 #include <opm/common/utility/platform_dependent/reenable_warnings.h>
+
 #include <opm/common/OpmLog/OpmLog.hpp>
 #include <opm/common/OpmLog/CounterLog.hpp>
 
 #include <opm/grid/CpGrid.hpp>
 #include <dune/grid/common/mcmgmapper.hh>
 
-#include <opm/core/props/satfunc/RelpermDiagnostics.hpp>
-#include <opm/input/eclipse/Parser/Parser.hpp>
-#include <opm/input/eclipse/Deck/Deck.hpp>
 #include <opm/input/eclipse/EclipseState/EclipseState.hpp>
 #include <opm/input/eclipse/EclipseState/Tables/TableManager.hpp>
+
+#include <opm/input/eclipse/Deck/Deck.hpp>
+
+#include <opm/input/eclipse/Parser/Parser.hpp>
 
 #if HAVE_DUNE_FEM
 #include <dune/fem/misc/mpimanager.hh>
